feat: migrated timer logic to timer.ts
