+++ src/main.js
@@ -1,29 +1,17 @@
-
-const startingTime = 30 * 60 * 100;
-let time = startingTime;
-let isTimerOn = false;
-let timerId = null;
-
+import {
+    toggleTimerLogic,
+    resetTimer,
+    timerOn,
+    getTime
+} from '../dist/timer.js';
+console.log('main.js loaded');
 const timer = document.getElementById('timer');
 const startBtn = document.getElementById('startBtn');
 const resetBtn = document.getElementById('resetBtn');
 const customiseBtn = document.getElementById('customiseBtn');
 
-function startTimer() {
-    if (isTimerOn === true) {
-        // Timer is counting down, pause it
-        clearInterval(timerId);
-        isTimerOn = false;
-        startBtn.innerHTML = 'Start';
-    } else {
-        updateTimer();
-        timerId = setInterval(updateTimer, 10);
-        isTimerOn = true;
-        startBtn.innerHTML = 'Pause';
-    }
-}
-
-function updateDisplay() {
+export function updateDisplay() {
+    const time = getTime();
     const totalSeconds = Math.floor(time / 100);
     let hours = Math.floor(totalSeconds / (60 * 60));
     let minutes = Math.floor((totalSeconds % (60 * 60)) / 60);
@@ -34,35 +22,18 @@ function updateDisplay() {
     seconds = seconds < 10 ? '0' + seconds : seconds;
 
     timer.innerHTML = `${hours}:${minutes}:${seconds}`;
+
+    startBtn.innerHTML = timerOn() ? 'Pause' : 'Start';
 }
 
-function updateTimer() {
-    if (time <= 0) {
-        clearInterval(timerId);
-        timer.innerHTML = '00:00:00:00';
-        time = startingTime;
-        isTimerOn = false;
-        return;
-    }
+setInterval(updateDisplay, 10);
 
-    updateDisplay();
-    time--;
-}
-
-function resetTimer() {
-    clearInterval(timerId);
-    time = startingTime;
-    isTimerOn = false;
-    startBtn.innerHTML = 'Start';
-
-    updateDisplay();
-}
 
 function customiseTimer() {
     resetTimer();
     // then begin adding functionality for customising the timer
 }
 
-startBtn.addEventListener('click', startTimer);
+startBtn.addEventListener('click', toggleTimerLogic);
 resetBtn.addEventListener('click', resetTimer);
 customiseBtn.addEventListener('click', customiseTimer);
