+++ src/timer.html
@@ -18,6 +18,6 @@
         </div>
     </div>
 
-    <script src="main.js"></script>
+    <script type="module" src="./main.js"></script>
 </body>
 </html>
