+++ tsconfig.json
@@ -1,14 +1,19 @@
 {
   "compilerOptions": {
+    "module": "ESNext",
+    "skipLibCheck": true,
+    "moduleResolution": "Node",
     "esModuleInterop": true,
-    "noImplicitAny": true,
-    "noEmit": true,
-    "resolveJsonModule": true,
-    "allowJs": true,
-    "target": "es2015",
-    "module": "commonjs"
+    "target": "ES2020",
+    "rootDir": "src",
+    "outDir": "dist",
   },
+  "include": [
+    "src/timer.ts"
+  ],
   "exclude": [
-    "**/**_broken.ts"
+    "**/**_broken.ts",
+    "eslint.config.mjs",
+    "jest.config.js",
   ],
 }
